+++ src/main/java/com/unrealdinnerbone/simplezoom/SimpleZoom.java
@@ -1,6 +1,7 @@
 package com.unrealdinnerbone.simplezoom;
 
 import net.minecraft.client.settings.KeyBinding;
+import net.minecraftforge.client.event.EntityViewRenderEvent;
 import net.minecraftforge.client.event.FOVUpdateEvent;
 import net.minecraftforge.common.config.Config;
 import net.minecraftforge.common.config.ConfigManager;
@@ -8,24 +9,19 @@ import net.minecraftforge.fml.client.event.ConfigChangedEvent;
 import net.minecraftforge.fml.client.registry.ClientRegistry;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
-import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import net.minecraftforge.fml.common.eventhandler.SubscribeEvent;
-import net.minecraftforge.fml.relauncher.Side;
 import org.lwjgl.input.Keyboard;
 
 @Mod.EventBusSubscriber
-@Mod(modid = SimpleZoom.MOD_ID, clientSideOnly = true)
+@Mod(modid = SimpleZoom.MOD_ID, clientSideOnly = true, version = SimpleZoom.MOD_VERSION, name = SimpleZoom.MOD_NAME)
 public class SimpleZoom {
 
+    public static final String MOD_NAME = "Simple Zoom";
     public static final String MOD_ID = "simplezoom";
+    public static final String MOD_VERSION = "@VERSION@";
 
     private static KeyBinding zoomBind;
 
-
-    @Mod.EventHandler
-    public static void onPreInit(FMLPreInitializationEvent event) {
-        ConfigManager.sync(MOD_ID, Config.Type.INSTANCE);
-    }
 
     @SubscribeEvent
     public static void onConfigChanged(ConfigChangedEvent.OnConfigChangedEvent event) {
@@ -36,24 +32,34 @@ public class SimpleZoom {
 
     @Mod.EventHandler
     public static void onInit(FMLInitializationEvent initializationEvent) {
-        if(initializationEvent.getSide() == Side.CLIENT) {
-            zoomBind = new KeyBinding("key.zoom.desc", Keyboard.KEY_Z, "key.zoom.catg");
-            ClientRegistry.registerKeyBinding(zoomBind);
+        zoomBind = new KeyBinding("key.zoom.desc", Keyboard.KEY_C, "key.zoom.catg");
+        ClientRegistry.registerKeyBinding(zoomBind);
+    }
+
+    @SubscribeEvent
+    public static void onFOVUpdateEvent(FOVUpdateEvent event) {
+        if (ZoomConfig.zoomType == ZoomType.FOV && zoomBind.isKeyDown()) {
+            event.setNewfov(-ZoomConfig.zoomAmount);
+        }
+    }
+    @SubscribeEvent
+    public static void onFOVModifierEvent(EntityViewRenderEvent.FOVModifier event) {
+        if (ZoomConfig.zoomType == ZoomType.CAMERA && zoomBind.isKeyDown()) {
+            event.setFOV(ZoomConfig.zoomAmount);
         }
     }
 
-    @SubscribeEvent()
-    public static void onEvent(FOVUpdateEvent event) {
-        if (zoomBind.isKeyDown()) {
-            event.setNewfov(ZoomConfig.zoomAmount);
-        }
-    }
-
-    @Config(modid = SimpleZoom.MOD_ID)
+    @Config(modid = SimpleZoom.MOD_ID, name = "../local/client/simplezoom")
     public static class ZoomConfig {
 
-        public static float zoomAmount = -5.0f;
+        public static float zoomAmount = 5.0f;
+        public static ZoomType zoomType = ZoomType.CAMERA;
 
+    }
+
+    public static enum ZoomType {
+        FOV,
+        CAMERA
     }
 
 }