+++ changelog.md
@@ -1,1 +1,3 @@
-+ Fixed Bug #2 Where Mod crashed due to class loading bug 
+###Version 1.2.0
+* Changed Location of config file to local/client/simplezoom
+* Cleaned up code thanks LatvainModder